pyproject.toml
[tool.poetry]
name = "fnbbmusicbot"
version = "0.1.0"
description = "Discord Bot specifically for streaming music."
authors = ["ethanlchristensen <ethanlchristensen@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.3.2"
python-dotenv = "^1.0.0"
youtube_dl = "^2021.12.17"
PyNaCl = "^1.5.0"
bruhcolor = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
